backend/requirements.txt

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0

# Async support
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# LLM and AI dependencies
google-generativeai==0.3.2
openai==1.3.7  # Alternative LLM option
anthropic==0.7.8  # Alternative LLM option

# Langraph for agentic workflows
langgraph==0.0.35
langchain==0.1.0
langchain-core==0.1.7

# Token counting and text processing
tiktoken==0.5.2
nltk==3.8.1
textstat==0.7.3

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy>=1.26.0


# Caching and memory
redis==4.6.0
# python-redis==4.6.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing dependencies
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# httpx-mock==0.21.0


# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.13.0
mypy==1.7.1

# Database (optional - for production)
sqlalchemy==2.0.23
alembic==1.12.1
# asyncpg==0.29.0  # PostgreSQL async driver

# CORS and middleware
# python-cors==1.7.0

# JSON and serialization
orjson==3.9.10

# Date and time utilities
python-dateutil==2.8.2

# Regex and text utilities
regex==2023.10.3

# Memory management
psutil==5.9.6

# Optional: For enhanced text analysis
spacy==3.7.2  # Advanced NLP
scikit-learn==1.3.2  # ML utilities for similarity calculations

# Optional: For document processing
python-docx==1.1.0
PyPDF2==3.0.1